+++ lib/scanner/vuln_scan.rb
@@ -15,6 +15,7 @@ module Yawast
 
         # server specific checks
         Yawast::Scanner::Plugins::Servers::Apache.check_all(uri)
+        Yawast::Scanner::Plugins::Servers::Nginx.check_all(uri)
         Yawast::Scanner::Plugins::Servers::Iis.check_all(uri, head)
 
         # checks for interesting files